Refactor gemm_afp4wfp4 configs
Refactor batched_gemm_afp4wfp4_pre_quant configs
